requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Deprecated==1.2.14
Django==3.2.9
django-debug-toolbar==3.6.0
django-jsonfield==1.4.1
django-redis-cache==3.0.1
djangorestframework==3.15.1
flake8==5.0.4
mccabe==0.7.0
packaging==24.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.1
pytz==2024.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
wrapt==1.16.0
